Updated the readme to add the note about pairing
BLE devices with security options take a few moments to initialise for the very first time. It's just on the very first pairing after you flash but then returns to the connection/pairing speeds one would expect from devices like BLE keyboards and mice. 

Yes it still happens if you set it to Just Works mode with no password, because the library has the ability to hotswap PIN security on and off so turning it off to begin with won't save you just given how I've designed this