Updated the Readme to make the SDL stuff clearer
When you pair the ESP32 with your computer after formatting, Dolphin, RPCS3, etc just pick up ALL inputs. It already knows that the hat = the D-pad, it already knows which stick is which and which axis is which. It just works, much like if you'd used a first-party PS3 or Xbox 360 controller with any popular emulator