NKRO/Gamepad swapping glitches fixed
Fixed the problem where if you repeatedly call the sendGamepadReport helper while mashing NKRO keys and holding down a modifier, things break

Kept tweaking tiny things and re-testing till it worked, idk why the changes I've made to the codebase fixed it. Still possible to trigger the glitch on mobile but it's VERY finicky and annoying to do on purpose now, plus it goes away never to return if you trigger a screen rotation one time. I'm calling it good enough as it's getting pretty difficult to run into this problem even if you're trying to do it intentionally